Change function name to `setWindowTitle`
On Linux at least, it's a bit weird for us to use "name" inconsistently between this function and `findByName`. We never actually change `WM_NAME` and it wouldn't have a visible effect if we did.

Note that we can't use `_NET_WM_NAME` for searching as it doesn't initially exist.