Add window icon (#8)
Add support for setting window icon (on Windows and Linux)

Useful resources for figuring out how to do this on Linux:
- https://tronche.com/gui/x/xlib/function-index.html
- https://specifications.freedesktop.org/wm-spec/wm-spec-1.3.html#idm45643490113536
- http://www.leonerd.org.uk/code/xseticon/ (studying implementation made me eventually realise I needed to call `flush` to see any effect)

We also implement support for setting the window's title, using similar mechanisms, and use this to avoid requiring the Gloss fork. In fact, we also now avoid setting the window title on every frame, instead only when we switch lights.