Safer integer conversions
Changes in GHC 9.2 mean that some `unsafeCoerce` calls which were once safe no longer are. I'm not even sure if we lose any performance.

The issue manifested as a weirdly discoloured logo in the window icon.

Note that the first such conversion, as opposed to the later four, wasn't necessary to get the logo to show properly again. But I've changed it anyway for consistency.